backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase>=2.6,<3
PyPDF2==3.0.1
openai==1.12.0
werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0